requirements.txt
# existing deps
openai>=1.0,<2.0
openai-whisper>=20231117
librosa>=0.10,<0.11
scipy>=1.11,<2.0
gTTS>=2.4,<3.0
tensorflow>=2.15,<3.0
scikit-learn>=1.3,<2.0
numpy>=1.24,<2.0
pandas>=2.0,<3.0
matplotlib>=3.7,<4.0
seaborn>=0.13,<1.0
gradio>=4.0,<5.0
tqdm>=4.66,<5.0
ipython>=8.0,<9.0

# Phase 1-3 additions
fastapi>=0.111,<1.0
uvicorn[standard]>=0.29,<1.0
pydantic>=2.6,<3.0
pydantic-settings>=2.2,<3.0
python-multipart>=0.0.9,<1.0
python-dotenv>=1.0,<2.0
redis>=5.0,<6.0
